requirements.txt
# Viriato - Python Dependencies
# Install with: pip install -r requirements.txt

# PostgreSQL adapter
psycopg2-binary>=2.9.9

# Flask API
Flask>=3.0.0
Flask-CORS>=4.0.0
gunicorn>=21.2.0  # Production WSGI server

# HTTP requests (for download_datasets.py)
requests>=2.31.0

# Optional: For development and testing
python-dotenv>=1.0.0  # Load DATABASE_URL from .env file
